fix: timer should switch session only on user interaction
The timer shouldn't switches state without user interaction; that is,
the timer shouldn't go from a work session to a break (or vice versa)
session without the user explicitly choosing to do so. This actively
engages the user to specify when they are working and when they are
taking a break. It also provides the user with the freedom of extending
the current work or break sessions.